backend/requirements.txt
djangorestframework
djangorestframework-jwt
django
djangorestframework-simplejwt
psycopg2
django-filter
django-cors-headers